Update company.py
Change the assignment of grid to find the element by the necessary classname (artdeco-card.p4.mb3) instead of the section number. This fixes the issue where a lot of fields were empty/null. But this isn’t very resilient to any changes LinkedIn might make to their classnames. I am not sure if it is better to use this approach or the numbering approach.